requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
-e git+https://github.com/sebatyler/py_translator@master#egg=py_translator
    # via -r requirements.in
apscheduler==3.6.3
    # via python-telegram-bot
argcomplete==2.0.0
    # via zappa
asttokens==2.4.1
    # via stack-data
attrs==22.2.0
    # via pytest
awslambdaric==2.0.4
    # via -r requirements.in
bcrypt==3.1.7
    # via paramiko
beautifulsoup4==4.6.3
    # via
    #   -r requirements.in
    #   django-bootstrap4
bepro-python-honey==0.0.1.6
    # via -r requirements.in
black==22.12.0
    # via -r requirements.in
boto3==1.26.45
    # via
    #   kappa
    #   zappa
botocore==1.29.45
    # via
    #   boto3
    #   s3transfer
build==0.9.0
    # via pip-tools
cachetools==4.2.2
    # via python-telegram-bot
certifi==2019.11.28
    # via
    #   python-telegram-bot
    #   requests
cffi==1.15.1
    # via
    #   bcrypt
    #   cryptography
    #   pynacl
cfn-flip==1.3.0
    # via troposphere
chardet==3.0.4
    # via requests
click==8.1.3
    # via
    #   black
    #   cfn-flip
    #   kappa
    #   pip-tools
cryptography==3.3.2
    # via paramiko
decorator==5.1.1
    # via ipython
django==2.2.28
    # via -r requirements.in
django-bootstrap4==1.1.0
    # via -r requirements.in
django-dotenv==1.4.2
    # via -r requirements.in
durationpy==0.5
    # via zappa
executing==2.0.1
    # via stack-data
fabric==2.7.1
    # via -r requirements.in
fabric3==1.14.post1
    # via -r requirements.in
gspread==3.1.0
    # via -r requirements.in
hjson==3.1.0
    # via zappa
httplib2==0.19.0
    # via oauth2client
idna==2.9
    # via requests
iniconfig==2.0.0
    # via pytest
invoke==1.7.3
    # via fabric
ipython==8.24.0
    # via -r requirements.in
isort==5.11.4
    # via -r requirements.in
jedi==0.19.1
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
    #   zappa
kappa==0.6.0
    # via zappa
langdetect==1.0.7
    # via py-translator
lxml==4.9.1
    # via py-translator
markupsafe==2.1.1
    # via werkzeug
matplotlib-inline==0.1.7
    # via ipython
mypy-extensions==0.4.3
    # via black
oauth2client==4.1.3
    # via -r requirements.in
packaging==21.3
    # via
    #   build
    #   pytest
paramiko==2.7.1
    # via
    #   fabric
    #   fabric3
parso==0.8.4
    # via jedi
pathlib2==2.3.7.post1
    # via fabric
pathspec==0.10.3
    # via black
pep517==0.13.0
    # via build
pexpect==4.9.0
    # via ipython
pip-tools==6.12.1
    # via -r requirements.in
placebo==0.9.0
    # via
    #   kappa
    #   zappa
platformdirs==2.6.2
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.43
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via oauth2client
pycparser==2.19
    # via cffi
pydash==5.1.1
    # via -r requirements.in
pygments==2.18.0
    # via ipython
pynacl==1.3.0
    # via paramiko
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
pytest==7.2.1
    # via
    #   pytest-django
    #   pytest-mock
pytest-django==4.5.2
    # via -r requirements.in
pytest-mock==3.10.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   zappa
python-slugify==4.0.0
    # via
    #   py-translator
    #   zappa
python-telegram-bot==13.15
    # via -r requirements.in
pytz==2019.3
    # via
    #   apscheduler
    #   django
    #   python-telegram-bot
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   cfn-flip
    #   kappa
    #   zappa
regex==2020.2.20
    # via somajo
requests==2.23.0
    # via
    #   gspread
    #   py-translator
    #   zappa
rsa==4.7
    # via oauth2client
s3transfer==0.6.0
    # via boto3
simplejson==3.17.2
    # via awslambdaric
six==1.14.0
    # via
    #   apscheduler
    #   asttokens
    #   bcrypt
    #   cfn-flip
    #   cryptography
    #   fabric3
    #   langdetect
    #   oauth2client
    #   pathlib2
    #   pynacl
    #   python-dateutil
somajo==1.8.3
    # via -r requirements.in
sqlparse==0.4.3
    # via django
stack-data==0.6.3
    # via ipython
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via zappa
tornado==6.1
    # via python-telegram-bot
tqdm==4.64.1
    # via zappa
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
troposphere==4.2.0
    # via zappa
typing-extensions==4.11.0
    # via ipython
tzdata==2022.7
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler
urllib3==1.25.8
    # via
    #   botocore
    #   requests
wcwidth==0.2.13
    # via prompt-toolkit
werkzeug==2.2.2
    # via zappa
wheel==0.38.4
    # via
    #   pip-tools
    #   zappa
zappa==0.56.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
